+++ src/com/euhedral/game/Entities/Enemy/EnemyStatic3.java
@@ -139,7 +139,7 @@ public class EnemyStatic3 extends Enemy {
     }
 
     @Override
-    public void resurrect(int x, int y) {
+    public void resurrect(double x, double y) {
         commonInit();
         explosion.playedOnce = false;
         super.resurrect(x, y);
